factory-optimizer/backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1

# ML
scikit-learn==1.3.2
xgboost==2.0.3
scipy==1.11.4
statsmodels==0.14.0

# LLM
openai>=1.0.0
anthropic>=0.7.0

# Utilities
python-dotenv==1.0.0
shap==0.43.0
joblib==1.3.2

